# Use index of encoding whether it need to convert