It can get the rb_encoding object based on its index value.

```c
rb_encoding*
rb_enc_get(VALUE obj)
{
    return rb_enc_from_index(rb_enc_get_index(obj));
}
```

Since UTF-8 has been used widely, it is sufficient to know only the index in most cases in this part.

−               | before | after  | result
--               | --     | --     | --
Oj.dump          | 1.026M | 1.119M | 1.091x

### Environment
- MacBook Pro (14 inch, 2021)
- macOS 12.0
- Apple M1 Max
- Ruby 3.1.0

### Before
```
Warming up --------------------------------------
             Oj.dump   102.823k i/100ms
Calculating -------------------------------------
             Oj.dump      1.026M (± 0.4%) i/s -     10.282M in  10.024252s
```

### After
```
Warming up --------------------------------------
             Oj.dump   110.736k i/100ms
Calculating -------------------------------------
             Oj.dump      1.119M (± 0.4%) i/s -     11.295M in  10.093770s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json =<<-EOF
{
  "$id": "https://example.com/person.schema.json",
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "title": "Person",
  "type": "object",
  "properties": {
    "firstName": {
      "type": "string",
      "description": "The person's first name."
    },
    "lastName": {
      "type": "string",
      "description": "The person's last name."
    },
    "age": {
      "description": "Age in years which must be equal to or greater than zero.",
      "type": "integer",
      "minimum": 0
    }
  }
}
EOF

Benchmark.ips do |x|
  x.warmup = 10
  x.time = 10

  data = Oj.load(json)
  x.report('Oj.dump') { Oj.dump(data, mode: :compat) }
end
```